+++ packages/calcite-components/src/custom-theme.stories.ts
@@ -100,6 +100,12 @@ import { meter, meterTokens } from "./custom-theme/meter";
 import { table, tableTokens } from "./custom-theme/table";
 import { carousel, carouselTokens } from "./custom-theme/carousel";
 import { dialog, dialogTokens } from "./custom-theme/dialog";
+import { swatchGroup, swatchGroupTokens } from "./custom-theme/swatch-group";
+import { swatch, swatchTokens } from "./custom-theme/swatch";
+import { splitButton, splitButtonTokens } from "./custom-theme/split-button";
+import { inputTimePicker, inputTimePickerTokens } from "./custom-theme/input-time-picker";
+import { colorPicker, colorPickerTokens } from "./custom-theme/color-picker";
+
 
 const globalTokens = {
   calciteColorBrand: "#007ac2",
@@ -245,6 +251,21 @@ const kitchenSink = (args: Record<string, string>, useTestValues = false) =>
     <div class="demo-row">
       <div class="demo-column">${dialog}</div>
     </div>
+    <div class="demo-row">
+      <div class="demo-column">${swatchGroup}</div>
+    </div>
+    <div class="demo-row">
+      <div class="demo-column">${swatch}</div>
+    </div>
+    <div class="demo-row">
+      <div class="demo-column">${splitButton}</div>
+    </div>
+    <div class="demo-row">
+      <div class="demo-column">${colorPicker}</div>
+    </div>
+    <div class="demo-row">
+      <div class="demo-column">${inputTimePicker}</div>
+    </div>
   </div>`;
 
 const componentTokens = {
@@ -314,6 +335,11 @@ const componentTokens = {
   ...tableTokens,
   ...carouselTokens,
   ...dialogTokens,
+  ...swatchGroupTokens,
+  ...swatchTokens,
+  ...splitButtonTokens,
+  ...inputTimePickerTokens,
+  ...colorPickerTokens,
 };
 
 export default {